+++ NEWS.md
@@ -1,3 +1,5 @@
+# squat (development version)
+
 # squat 0.3.0
 
 ## New features